backend/requirements.txt
aiofiles==24.1.0
aiohttp==3.12.14
fastapi==0.116.1
python-jose==3.3.0
passlib==1.7.4
pydantic==2.11.7
pydantic_settings==2.10.1
SQLAlchemy==2.0.41
uvicorn==0.35.0
bcrypt==4.0.1
python-multipart==0.0.20
